Added dynamic data loading to page analytics
ref https://linear.app/ghost/issue/PROD-234/build-post-analytics-skeleton

- So far the React based post analytics skeleton was completely static. This PR added reading post and traffic data which makes it actually usable to beta users.
