+++ apps/shade/package.json
@@ -50,6 +50,7 @@
         "jsdom": "24.1.3",
         "lodash-es": "4.17.21",
         "mocha": "10.8.2",
+        "moment-timezone": "^0.5.48",
         "react": "18.3.1",
         "react-dom": "18.3.1",
         "rollup-plugin-node-builtins": "2.1.2",
